Fix formatting issue in number_format Rd
The use of `%` in the details is confusing roxygen. If you view the docs using `?number_format`, everything on the line after `%` in details is omitted because it is being interpreted like a LaTeX comment. I think escaping them should fix it. I haven't searched other parts of the docs, so there could be other instances.